requirements.txt
pyTelegramBotAPI==4.12.0
python-dotenv
SQLAlchemy
